[FIX_FOR_VLLM_CUSTOM=d28d86e8a34bf2617be294c235d6e6ef3321917b] (#1279)
1) KV cache offloading (vllm#37853):
cpu_hpu.py: Rewrite HPU monkey-patches for new CanonicalKVCaches API —
updated get_handlers, SingleDirectionOffloadingHandler.__init__,
CpuGpuOffloadingHandlers.__init__, and transfer_async. Added
OffloadingConnectorWorker.register_kv_caches override to handle HPU's
TensorTuple (K/V pair) KV cache layout.
test_offloading_connector.py: Renamed to utils.py per upstream
restructuring.
offloading_connector: New test subdirectory with __init__.py,
conftest.py, utils.py, test_scheduler.py, test_metrics.py matching
upstream test split.

2) GDN attention refactor (vllm#37975):
qwen3_5.py: Updated monkey-patch target from removed
Qwen3_5GatedDeltaNet to GatedDeltaNetAttention.

3) DeepSeek OCR fix (vllm#35182):
deepseek_ocr.py: Minor compatibility fix.

4) Pooling metadata (vllm#38139):
hpu_input_batch.py: Renamed _make_prompt_token_ids_tensor →
_make_prompt_token_ids_cpu_tensor (returns CPU tensor); updated
_make_sampling_metadata to create CPU tensor first then copy to device;
added prompt_token_ids_cpu to get_pooling_metadata.
hpu_model_runner.py: Added prompt_token_ids_cpu argument to
PoolingMetadata constructors.

5) LLM(**engine_args)  test changes (vllm#37902):
generation_mm.py — replace asdict(engine_args) + LLM(**engine_args) with
LLM.from_engine_args(engine_args), remove unused asdict import
generation_mm_multi.py — same change

---------
