+++ requirements.txt
@@ -5,3 +5,4 @@ numba>=0.58.0
 numpy>=1.26.0
 transformers>= 4.56.0, <5
 kaldi-native-fbank >= 1.18.7
+tblib==3.1.0
